Make links within conference tabs load whole page
This fixes the bug where internal links within page parts were not correctly
loading. Unfortunately, it somewhat negates the point of using turbo
frames in the first place, since it also forces the tab links themselves
to load the whole page. It's worth looking for a better solution for
this in future.